Reverted schema files back to master state.
