+++ src/js/pagi.js
@@ -1,6 +1,7 @@
 var GraphParser = require('./graph/parsers/graphParser');
-var SchemaParser = require('./schema/schemaParser');
 
 module.exports = function() {
   return "I AM PAGI";
 };
+
+module.exports.schema = require('./schema');
